# docs/api/openapi.draft.yaml
openapi: 3.1.0
info:
  title: Purrenade API
  version: 0.1.0-draft
  summary: Contract shape for the Purrenade backend.
  description: |
    **M0 DRAFT — contract shape only. No implementation exists.**

    This document defines the intended shape of the Purrenade REST API so the
    frontend can be designed against it before the backend is bootstrapped. It is
    not generated from code, and no controller, migration, or endpoint exists yet.

    Authoritative product behavior lives in the frontend repository's
    Product/Game Specification (`purrenade/docs/product/`). Conventions for
    versioning, errors, pagination, idempotency and rate limiting are in
    `docs/api/api-conventions.md`.

    **Decided in M0.5:**

    - **ADR-0005 — Accepted.** A Nuxt BFF with server-managed session cookies for the
      browser, over a **token-capable** API. No endpoint may assume a browser, a cookie,
      or a same-site context: a future native client authenticates directly against this
      API with a bearer flow, bypassing the BFF.
    - **Achievement progression authority — APPROVED.** Client-reported summary counters
      alone are insufficient. Telemetry fields in `FinishRunRequest` are **inputs to
      server-side derivation, never authoritative totals**. Each achievement carries a
      **Verification Source** and a **Progress Persistence**; storing a cumulative total
      never reclassifies where its evidence came from.
...
      not thresholds earned.
...

    **Still open:**

...

...
servers:
...
security:
  - sessionToken: []
paths:
...
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: End this session
      description: |
        **IMPLEMENTED (M2).** Deletes the token row, so the credential stops existing —
        revocation is real, not a cookie being forgotten.

        This session only. Other devices are untouched; `DELETE /auth/sessions` is the
        endpoint for that intent.

        Reachable with an unverified account: a player must always be able to leave.
      responses:
        '204':
          description: The session has ended.
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        **IMPLEMENTED (M2).** A Sanctum personal access token, presented as
        `Authorization: Bearer <token>`. One token is one signed-in device — the rows it
        addresses are what `/auth/sessions` lists and revokes.

        **Two callers, one scheme, and that is the point (ADR-0005):**

        - **Browser.** The token is issued to the **Nuxt BFF**, which keeps it in its own
          server-side session and attaches it upstream. It never reaches browser
          JavaScript, and no bearer token is stored in `localStorage` or `sessionStorage`.
          The browser's only credential is an `HttpOnly`, `Secure`, `SameSite=Lax` cookie
          set by the BFF, which this API never sees.

        - **Future native client.** Authenticates directly against these same endpoints
          with the same scheme, bypassing the BFF. Not implemented in v1; the obligation
          it places on this contract is that **no endpoint may assume a browser, a cookie,
          or a same-site context**.

        **Abilities record how the session was authenticated.** Every token carries
        `session`; only a token minted by `POST /auth/2fa/challenge` also carries
        `two-factor`, and no endpoint adds that ability to an existing token. Admin-scoped
        routes require it, which is how "this session passed a second factor" is
        distinguished from "this account has a second factor". Tokens are never issued
        with the `*` wildcard.

        **CSRF is not a concern on this hop.** The BFF is not a browser and attaches the
        credential explicitly, so no ambient authority exists here. CSRF is enforced at the
        BFF, where the cookie is.

        Laravel remains the authentication authority. The BFF carries credentials; it makes
        no authorization decision.
  responses:
    Unauthenticated:
      description: |
        No credential, or one that no longer resolves. `code`: `unauthenticated`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      description: |
        **The** error shape. RFC 9457 Problem Details, served as
        `application/problem+json`, for every error this API produces — validation,
        authentication, authorization, conflict, rate limiting, and unexpected failures
        alike (API-1).

        One shape means a client needs one parser. There is no second envelope anywhere in
        this contract.

        **Never present:** stack traces, SQL, internal class names, file paths, dependency
        versions. That holds in every environment, including with debug mode on — a
        response shape that changes between environments is one nobody can test against.
      properties:
        type:
          type: string
          format: uri
          description: |
            A URN identifying the problem type, always `urn:purrenade:error:{code}`.

            A URN rather than an `https://` URL: RFC 9457 does not require the type to be
            dereferenceable, and no published documentation site exists to dereference.
            Inventing a URL that returns 404 would be worse than being honest.
          examples: ['urn:purrenade:error:validation_failed']
        title:
          type: string
          description: |
            A short, stable, **English** summary of the problem *type*, not of this
            occurrence. Not localised — see `detail`.
        status:
          type: integer
          description: The HTTP status, repeated in the body as RFC 9457 specifies.
          examples: [422]
        detail:
          type: string
          description: |
            Human-readable explanation of **this occurrence**. For display and logs only.
            May be reworded or localised at any time, so **never branch on it**.
        instance:
          type: string
          description: The path that produced the problem.
          examples: ['/api/v1/auth/register']
        code:
          type: string
          description: |
            **The value clients branch on.** Stable, machine-readable, never localised.
            Always equal to the tail of `type`, so the two cannot disagree.

            Adding a code is additive and safe. Changing or removing one is breaking and
            requires a new API version.
          examples: ['validation_failed']
        correlation_id:
          type: string
          description: |
            Present on every problem, and echoed in the `X-Correlation-Id` response header.
            The player can quote it; it appears in this service's logs and in the BFF's.
        errors:
          type: object
          description: |
            Present on validation failures. Maps a field name to the list of things wrong
            with it, each with its own stable code — so a client can render "already taken"
            in Turkish without parsing an English sentence.
          additionalProperties:
            type: array
            items:
              type: object
              properties:
                code:
                  type: string
                  description: |
                    Stable per-field code, e.g. `required`, `taken`, `too_short`,
                    `too_long`, `email_invalid`, `confirmation_mismatch`,
                    `password_incorrect`, `password_compromised`, `format_invalid`.
                message:
                  type: string
                  description: Human-readable. For display only.
              required: [code, message]
        retry_after:
          type: integer
          minimum: 0
          description: |
            Present on `429`. Seconds to wait. Also sent as a `Retry-After` header — the
            header is what proxies and HTTP clients honour, the member is what a countdown
            in the UI needs.
      required: [type, title, status, detail, instance, code, correlation_id]
      examples:
        - type: urn:purrenade:error:validation_failed
          title: Validation failed
          status: 422
          detail: The given data was invalid.
          instance: /api/v1/auth/register
          code: validation_failed
          correlation_id: 01K5R8Q2M3N4P5Q6R7S8T9V0W1
          errors:
            email:
              - code: taken
                message: This email address is already registered.
...